Finished autocomplete functionality, moved Location Filter containers to appropriate folders
